.tools/requirements.txt
ruamel.yaml>=0.17.0
flake8
flake8-bugbear
pytest
pytest-cov
click